requirements.txt
streamlit
torch
torchvision
timm
opencv-python-headless
numpy
Pillow
matplotlib
huggingface_hub
fpdf2